requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
gunicorn==21.2.0
dnspython==2.4.2
python-telegram-bot==20.7
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
google-generativeai==0.3.2
# RAG System (from original project)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0